requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
einops==0.2.0
filelock==3.7.1
Flask==2.1.2
Flask-Cors==3.0.10
fonttools==4.34.2
gdown==4.4.0
idna==3.3
imageio==2.19.3
imgaug==0.4.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.3
lmdb==1.0.0
MarkupSafe==2.1.1
matplotlib==3.5.2
networkx==2.8.4
numpy==1.23.0
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
prefetch-generator==1.0.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.8.1
Shapely==1.8.2
six==1.16.0
soupsieve==2.3.2.post1
tifffile==2022.5.4
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.9
vietocr==0.3.8
Werkzeug==2.1.2
zipp==3.8.0